fixes #21

+++ src/lib/utils/pointcloud.py
@@ -279,10 +279,13 @@ def pc_dep_to_hm_torch(pc_hm, pc_dep, dep, bbox, dist_thresh, opt):
       dep = dep[0]
     ct = torch.tensor(
       [(bbox[0] + bbox[2]) / 2, (bbox[1] + bbox[3]) / 2], dtype=torch.float32)
-    bbox_int = torch.tensor([torch.floor(bbox[0]), 
-                         torch.floor(bbox[1]), 
-                         torch.ceil(bbox[2]), 
-                         torch.ceil(bbox[3])], dtype=torch.int32)# format: xyxy
+    # print("type bbox: ", type(bbox))
+    # print("bbox[0]: ", bbox[0])
+    # print("type(bbox[0]): ", type(bbox[0]))
+    bbox_int = torch.tensor([int(torch.floor(bbox[0])), 
+                         int(torch.floor(bbox[1])), 
+                         int(torch.ceil(bbox[2])), 
+                         int(torch.ceil(bbox[3]))], dtype=torch.int32)# format: xyxy
 
     roi = pc_dep[:, bbox_int[1]:bbox_int[3]+1, bbox_int[0]:bbox_int[2]+1]
     pc_dep = roi[opt.pc_feat_channels['pc_dep']]
